Move static serving to next server
