+++ packages/next/types/index.d.ts
@@ -75,6 +75,7 @@ export type PageConfig = {
      */
     externalResolver?: true
   }
+  runtime?: 'nodejs' | 'edge'
   env?: Array<string>
   unstable_runtimeJS?: false
   unstable_JsPreload?: false
